+++ integrations/streamer_test.go
@@ -197,6 +197,51 @@ func TestStream_ChangeDetect(t *testing.T) {
 
 	testStreamerWithOutput(t, "TestStream_ChangeDetect", script, 15*time.Second, er, false, nil)
 }
+func TestStream_ChangeDetect_Many(t *testing.T) {
+
+	var script = `stream
+	|from().measurement('packets')
+	|changeDetect('a','b')
+    |window()
+		.period(6s)
+		.every(6s)
+	|httpOut('TestStream_ChangeDetect_Many')
+`
+
+	er := models.Result{
+		Series: models.Rows{
+			{
+				Name:    "packets",
+				Tags:    nil,
+				Columns: []string{"time", "a", "b"},
+				Values: [][]interface{}{
+					[]interface{}{
+						time.Date(1971, 1, 1, 0, 0, 0, 0, time.UTC),
+						"bad",
+						0.0,
+					},
+					[]interface{}{
+						time.Date(1971, 1, 1, 0, 0, 1, 0, time.UTC),
+						"good",
+						0.0,
+					},
+					[]interface{}{
+						time.Date(1971, 1, 1, 0, 0, 4, 0, time.UTC),
+						"bad",
+						1.0,
+					},
+					[]interface{}{
+						time.Date(1971, 1, 1, 0, 0, 5, 0, time.UTC),
+						"bad",
+						0.0,
+					},
+				},
+			},
+		},
+	}
+
+	testStreamerWithOutput(t, "TestStream_ChangeDetect_Many", script, 15*time.Second, er, false, nil)
+}
 
 func TestStream_Derivative(t *testing.T) {
 
@@ -10071,9 +10116,9 @@ Value: {{ index .Fields "count" }}
 				"Mime-Version":              []string{"1.0"},
 				"Content-Type":              []string{"text/html; charset=UTF-8"},
 				"Content-Transfer-Encoding": []string{"quoted-printable"},
-				"To":      []string{"user1@example.com, user2@example.com"},
-				"From":    []string{"test@example.com"},
-				"Subject": []string{"kapacitor.cpu.serverA is CRITICAL"},
+				"To":                        []string{"user1@example.com, user2@example.com"},
+				"From":                      []string{"test@example.com"},
+				"Subject":                   []string{"kapacitor.cpu.serverA is CRITICAL"},
 			},
 			Body: `
 <b>kapacitor.cpu.serverA is CRITICAL</b>
@@ -10087,9 +10132,9 @@ Value: 10
 				"Mime-Version":              []string{"1.0"},
 				"Content-Type":              []string{"text/html; charset=UTF-8"},
 				"Content-Transfer-Encoding": []string{"quoted-printable"},
-				"To":      []string{"user1@example.com, user2@example.com"},
-				"From":    []string{"test@example.com"},
-				"Subject": []string{"kapacitor.cpu.serverA is CRITICAL"},
+				"To":                        []string{"user1@example.com, user2@example.com"},
+				"From":                      []string{"test@example.com"},
+				"Subject":                   []string{"kapacitor.cpu.serverA is CRITICAL"},
 			},
 			Body: `
 <b>kapacitor.cpu.serverA is CRITICAL</b>